Conflicts:
	flight/PiOS/Common/pios_rfm22b.c

+++ flight/Modules/System/systemmod.c
@@ -460,7 +460,7 @@ static void updateStats()
 	idleCounterClear = 1;
 	
 #if defined(PIOS_INCLUDE_ADC) && defined(PIOS_ADC_USE_TEMP_SENSOR)
-	float temp_voltage = 3.3 * PIOS_ADC_PinGet(0) / ((1 << 12) - 1);
+	float temp_voltage = 3.3f * PIOS_ADC_PinGet(0) / ((float)((1 << 12) - 1));
 	const float STM32_TEMP_V25 = 1.43; /* V */
 	const float STM32_TEMP_AVG_SLOPE = 4.3; /* mV/C */
 	stats.CPUTemp = (temp_voltage-STM32_TEMP_V25) * 1000 / STM32_TEMP_AVG_SLOPE + 25;
